requirements.txt
fFlask==3.0.0
Flask-Cors==4.0.0
python-dotenv==1.0.0
bcrypt==4.1.2
